Refactored Backend. Added some auth middleware
