+++ .github/workflows/backend-CI.yml
@@ -17,6 +17,8 @@ jobs:
       CI: true
       AUTH0_CLIENT_ID: ${{ secrets.AUTH0_CLIENT_ID }}
       AUTH0_CLIENT_SECRET: ${{ secrets.AUTH0_CLIENT_SECRET }}
+      AUTH0_PRIVILEGED_CLIENT_ID: ${{ secrets.AUTH0_PRIVILEGED_CLIENT_ID }}
+      AUTH0_PRIVILEGED_CLIENT_SECRET: ${{ secrets.AUTH0_PRIVILEGED_CLIENT_SECRET }}
       AUTH0_AUDIENCE: ${{ secrets.AUTH0_AUDIENCE }}
     steps:
       - uses: actions/checkout@v4
