+++ Backend/package.json
@@ -1,50 +1,50 @@
 {
-    "name": "backend",
-    "version": "1.0.0",
-    "description": "",
-    "license": "ISC",
-    "author": "",
-    "type": "commonjs",
-    "main": "backend.js",
-    "jest": {
-        "setupFilesAfterEnv": [
-            "./tests/setup/jest.setup.js",
-            "dotenv/config",
-            "./tests/setup/auth.setup.js"
-        ],
-        "coverageReporters": [
-            "lcov",
-            "text-summary"
-        ],
-        "reporters": [
-            "default",
-            [
-                "jest-junit",
-                {
-                    "outputDirectory": "coverage/",
-                    "outputName": "results.xml"
-                }
-            ]
-        ]
-    },
-    "scripts": {
-        "test": "jest unit",
-        "testall": "jest",
-        "testcov": "jest --coverage"
-    },
-    "dependencies": {
-        "cors": "^2.8.5",
-        "express": "^5.1.0",
-        "express-oauth2-jwt-bearer": "^1.6.1",
-        "morgan": "^1.10.0",
-        "pg": "^8.16.0",
-        "winston": "^3.17.0"
-    },
-    "devDependencies": {
-        "dotenv": "^16.5.0",
-        "jest": "^29.7.0",
-        "jest-junit": "^16.0.0",
-        "supertest": "^7.1.1",
-        "wait-on": "^8.0.3"
-    }
+	"name": "backend",
+	"version": "1.0.0",
+	"description": "",
+	"license": "ISC",
+	"author": "",
+	"type": "commonjs",
+	"main": "backend.js",
+	"jest": {
+		"setupFilesAfterEnv": [
+			"./tests/setup/jest.setup.js",
+			"dotenv/config",
+			"./tests/setup/auth.setup.js"
+		],
+		"coverageReporters": [
+			"lcov",
+			"text-summary"
+		],
+		"reporters": [
+			"default",
+			[
+				"jest-junit",
+				{
+					"outputDirectory": "coverage/",
+					"outputName": "results.xml"
+				}
+			]
+		]
+	},
+	"scripts": {
+		"test": "jest unit",
+		"testall": "jest",
+		"testcov": "jest --coverage"
+	},
+	"dependencies": {
+		"cors": "^2.8.5",
+		"express": "^5.1.0",
+		"express-oauth2-jwt-bearer": "^1.6.1",
+		"morgan": "^1.10.0",
+		"pg": "^8.16.0",
+		"winston": "^3.17.0"
+	},
+	"devDependencies": {
+		"dotenv": "^16.5.0",
+		"jest": "^29.7.0",
+		"jest-junit": "^16.0.0",
+		"supertest": "^7.1.1",
+		"wait-on": "^8.0.3"
+	}
 }
