+++ Backend/tests/setup/jest.setup.js
@@ -1,11 +1,11 @@
 const winston = require("winston");
 
 jest.spyOn(winston, "createLogger").mockImplementation(() => ({
-    log: jest.fn(),
-    info: jest.fn(),
-    error: jest.fn(),
-    warn: jest.fn(),
-    debug: jest.fn(),
-    clear: jest.fn(),
-    add: jest.fn(),
+	log: jest.fn(),
+	info: jest.fn(),
+	error: jest.fn(),
+	warn: jest.fn(),
+	debug: jest.fn(),
+	clear: jest.fn(),
+	add: jest.fn(),
 }));
